+++ main/src/de/fourtytwoways/onion/ContractDemo.java
@@ -10,6 +10,7 @@ import de.fourtytwoways.onion.infrastructure.enums.provider.ExampleEnumRepositor
 import de.fourtytwoways.onion.domain.model.enums.EnumType;
 import de.fourtytwoways.onion.domain.model.enums.Product;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class ContractDemo {
@@ -36,7 +37,7 @@ public class ContractDemo {
                 myContractRepository.createContract("42",
                         gemischteVersicherung,
                         LocalDate.of(2022, 4, 1), LocalDate.of(2042, 3, 31),
-                        666.);
+                        BigDecimal.valueOf( 666.));
         System.out.println(c1);
         myContractRepository.saveContract(c1);
 
@@ -54,7 +55,7 @@ public class ContractDemo {
                 myContractRepository.createContract("0815",
                         gemischteVersicherung,
                         LocalDate.of(2022, 4, 1), LocalDate.of(2042, 3, 31),
-                        47.11);
+                        BigDecimal.valueOf(47.11));
         System.out.println(c4);
         myContractRepository.saveContract(c4);
